+++ projetoIntegradorApi/src/main/java/com/example/projetoIntegrador/repository/ConsultaRepository.java
@@ -3,5 +3,5 @@ package com.example.projetoIntegrador.repository;
 import com.example.projetoIntegrador.model.Consulta;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ConsultaRepository extends JpaRepository<Consulta, Long> {
+public interface ConsultaRepository extends JpaRepository<Consulta, Integer> {
 }
