+++ projetoIntegradorApi/src/main/java/com/example/projetoIntegrador/repository/FarmaPacienteRepository.java
@@ -3,5 +3,5 @@ package com.example.projetoIntegrador.repository;
 import com.example.projetoIntegrador.model.FarmaPaciente;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FarmaPacienteRepository extends JpaRepository<FarmaPaciente, Long> {
+public interface FarmaPacienteRepository extends JpaRepository<FarmaPaciente, Integer> {
 }
